# feat(tensor): add BitLinear layer and tests (#178)