## Test Coverage
- Current coverage: 84.1%
- Coverage changes: 83.0% → 84.1%

## Performance Metrics
### Memory Usage
#### Tensor Operations
- Allocations per operation:
  - New tensor creation: 2 allocs/op (100, 100x100, 50x50x50, 20x20x20x20)
  - Get/Set operations: 0 allocs/op (2D access)
  - Parallel operations: 10,022 allocs/op (100x100), 1,000,023 allocs/op (1000x1000)

#### BitNet Model Operations
- Allocations per operation:
  - Model weights loading: 22 allocs/op (small), 22 allocs/op (medium), 22 allocs/op (large)
  - Model inference: N/A allocs/op (TODO #190)
  - Ternary weights reading: 1 allocs/op (small), 1 allocs/op (medium), 1 allocs/op (large)

### CPU Performance
#### Tensor Operations
- Operation timing:
  - Basic operations: 11.8 ns/op (Get), 10.7 ns/op (Set)
  - Parallel operations: 93,694 ns/op (100x100), 6,507,018 ns/op (1000x1000)
  - Large tensor operations: 1,238 ns/op (NewTensor 100x100), 13,093 ns/op (NewTensor 50x50x50), 16,230 ns/op (NewTensor 20x20x20x20)

#### BitNet Model Operations
- Operation timing:
  - Model weights loading: 417,682 ns/op (small), 1,578,995 ns/op (medium), 5,973,249 ns/op (large)
  - Model inference: N/A ns/op (TODO #190)
  - Ternary weights reading: 30,791 ns/op (small), 124,614 ns/op (medium), 216,101 ns/op (large)

## Areas for Improvement
### High Priority
- [ ] Optimize memory allocations in model operations (TODO #191)
- [ ] Implement proper self-attention (TODO #186)

### Medium Priority
- [ ] Improve error handling in model operations (TODO #192)
- [ ] Add more comprehensive benchmarks (TODO #192)
- [ ] Enhance documentation
- [ ] Implement proper feed-forward network (TODO #187)

### Low Priority
- [ ] Consider SIMD optimizations (TODO #191)
- [ ] Add more model operations (TODO #190)
- [ ] Improve test organization (TODO #192)
- [ ] Implement proper output generation (TODO #189)

Closes #178